refactor: migrate to wear-os-sensors lib features
